chore: Moves CronPicker to the components folder (#14125)
