+++ superset-frontend/src/common/components/common.stories.tsx
@@ -20,6 +20,7 @@ import React, { useState, useRef, useCallback } from 'react';
 import { action } from '@storybook/addon-actions';
 import { withKnobs, boolean, select } from '@storybook/addon-knobs';
 import Button from 'src/components/Button';
+import { CronPicker, CronError } from 'src/components/CronPicker';
 import Modal from './Modal';
 import Tabs, { EditableTabs } from './Tabs';
 import { Tooltip as AntdTooltip } from './Tooltip';
@@ -27,7 +28,6 @@ import { Switch as AntdSwitch } from './Switch';
 import { Menu, Input, Divider } from '.';
 import { Dropdown } from './Dropdown';
 import InfoTooltip from './InfoTooltip';
-import { CronPicker, CronError } from './CronPicker';
 
 export default {
   title: 'Common components',
